requirements/base.txt
pandas>=2.2.0
matplotlib>=3.9.0
numpy>=1.26.0
click>=8.1.0
pyyaml>=6.0.1
jinja2>=3.1.0
python-box>=7.0.0
requests>=2.32.0
